foauth/requirements.txt
requests==2.21.0
Flask==1.0.2
oauthlib==3.0.1
pyOpenSSL==19.0.0
Flask-Login==0.4.1
python-dotenv==0.10.3
facebook-sdk==3.1.0
